refactor: ResetButtonコンポーネントのパスを更新 ディレクトリの変更
